requirements.txt
Flask==0.10.1
itsdangerous==0.23
Jinja2==2.7.2
Pillow==2.6.1
Werkzeug==0.9.4
